Replace DrawRect rendering implementation with a naiive individual pixel renderer. This is MUCH slower for some reason...but I need the individual pixel access

+++ src/render/rgb.rs
@@ -1,9 +1,9 @@
 use std::ops;
 
 pub struct RGB {
-    pub red: u32,
-    pub green: u32,
-    pub blue: u32,
+    pub red: u8,
+    pub green: u8,
+    pub blue: u8,
 }
 
 impl ops::Div<f64> for RGB {
@@ -11,9 +11,9 @@ impl ops::Div<f64> for RGB {
 
     fn div(self, _rhs: f64) -> RGB {
         RGB {
-            red: (self.red as f64 / _rhs) as u32,
-            green: (self.green as f64 / _rhs) as u32,
-            blue: (self.blue as f64 / _rhs) as u32,
+            red: (self.red as f64 / _rhs) as u8,
+            green: (self.green as f64 / _rhs) as u8,
+            blue: (self.blue as f64 / _rhs) as u8,
         }
     }
 }
